T1675 FIX bad implementation
